% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpolypath-package.r
\docType{data}
\name{dathome}
\alias{dathome}
\alias{maphome}
\title{home data}
\description{
a home of three objects with multiple parts as two related data frames
}
\details{
\code{dathome} is the metadata
\code{maphome} is the geometry
}

