% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpolypath-package.r
\docType{data}
\name{gardenstate}
\alias{gardenstate}
\title{Province polygons with inland waters as holes.}
\description{
A data frame of coordinates and geometry classifiers
of the garden state, South Australia.
}
\details{
The PROJ.4 string for this map is:
'+proj=lcc +lat_1=-47 +lat_2=-17 +lat_0=-32 +lon_0=136 +x_0=0 +y_0=0
 +ellps=WGS84 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs'
}
\examples{
gs <- ggplot(gardenstate)
gs <- gs + aes(x = x, y = y, group = group, fill = id)
gs +  geom_polypath() + geom_path()

}

