% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpolypath-package.R
\docType{package}
\name{ggpolypath-package}
\alias{ggpolypath}
\alias{ggpolypath-package}
\title{ggpolypath: Polygons with Holes for the Grammar of Graphics}
\description{
Tools for working with polygons with holes in 'ggplot2', with a new 'geom' for drawing a 'polypath' applying the 'evenodd' or 'winding' rules.

Tools for working with polygons with holes in 'ggplot2', with a new 'geom' for drawing a 'polypath' applying the 'evenodd' or 'winding' rules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mdsumner.github.io/ggpolypath/}
  \item \url{http://rpubs.com/kohske/3522/}
  \item Report bugs at \url{https://github.com/mdsumner/ggpolypath/issues/}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Kohske Takahashi \email{takahashi.kohske@gmail.com} (original author of 'geom_holygon') [contributor]
}

}
\keyword{internal}
