% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpolypath-package.R
\docType{package}
\name{ggpolypath}
\alias{ggpolypath}
\title{ggpolypath}
\description{
Tools for working with polygons with holes in 'ggplot2', with a new 'geom' for drawing a 'polypath' applying the 'evenodd' or 'winding' rules.
}
